fix: align Feishu HITL guidance and split API tests
Render Feishu card answer instructions from answerTransport so respond
parks are not sent to GitHub and free-text is not described as resuming.
Split oversized handler_hitl_test.go into respond, Feishu card, and
thread contract files under the 300-line-added threshold.

Generated-By: looper 0.11.2 (runner=fixer, agent=grok-build)
